Remove calculate_stream_duration, add ffprobe to tests

+++ ffmpeg/ffmpeg.go
@@ -246,7 +246,7 @@ type MediaFormatInfo struct {
 	PixFormat      PixelFormat
 	Width, Height  int
 	FPS            float32
-	DurSecs        float32
+	DurSecs        int64
 }
 
 func (f *MediaFormatInfo) ScaledHeight(width int) int {
@@ -280,7 +280,7 @@ func GetCodecInfo(fname string) (CodecStatus, MediaFormatInfo, error) {
 	format.Width = int(params_c.width)
 	format.Height = int(params_c.height)
 	format.FPS = float32(params_c.fps)
-	format.DurSecs = float32(params_c.dur)
+	format.DurSecs = int64(params_c.dur)
 	return status, format, nil
 }
